[plugman] renaming GCM/ to gcm/ to make plugman work
